move snowflake configs to separate file
