+++ metadata-ingestion/src/datahub/ingestion/source/sql/snowflake.py
@@ -4,40 +4,28 @@ from collections import defaultdict
 from dataclasses import dataclass, field
 from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Union
 
-import pydantic
-
 # This import verifies that the dependencies are available.
 import snowflake.sqlalchemy  # noqa: F401
 import sqlalchemy.engine
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives import serialization
-from snowflake.connector.network import (
-    DEFAULT_AUTHENTICATOR,
-    EXTERNAL_BROWSER_AUTHENTICATOR,
-    KEY_PAIR_AUTHENTICATOR,
-)
 from snowflake.sqlalchemy import custom_types, snowdialect
 from sqlalchemy import create_engine, inspect
 from sqlalchemy.engine.reflection import Inspector
 from sqlalchemy.sql import sqltypes, text
 
 import datahub.emitter.mce_builder as builder
-from datahub.configuration.common import AllowDenyPattern
-from datahub.configuration.time_window_config import BaseTimeWindowConfig
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
 from datahub.ingestion.api.common import PipelineContext
 from datahub.ingestion.api.workunit import MetadataWorkUnit
 from datahub.ingestion.source.aws.s3_util import make_s3_urn
 from datahub.ingestion.source.sql.sql_common import (
     RecordTypeClass,
-    SQLAlchemyConfig,
     SQLAlchemySource,
     SQLSourceReport,
     SqlWorkUnit,
     TimeTypeClass,
-    make_sqlalchemy_uri,
     register_custom_type,
 )
+from datahub.ingestion.source_config.sql.snowflake import SnowflakeConfig
 from datahub.metadata.com.linkedin.pegasus2avro.dataset import (
     DatasetLineageTypeClass,
     UpstreamClass,
@@ -54,8 +42,6 @@ register_custom_type(custom_types.VARIANT, RecordTypeClass)
 
 logger: logging.Logger = logging.getLogger(__name__)
 
-APPLICATION_NAME = "acryl_datahub"
-
 snowdialect.ischema_names["GEOGRAPHY"] = sqltypes.NullType
 
 
@@ -70,128 +56,6 @@ class SnowflakeReport(SQLSourceReport):
     saas_version: str = ""
     role: str = ""
     role_grants: List[str] = field(default_factory=list)
-
-
-class BaseSnowflakeConfig(BaseTimeWindowConfig):
-    # Note: this config model is also used by the snowflake-usage source.
-
-    scheme = "snowflake"
-
-    username: Optional[str] = None
-    password: Optional[pydantic.SecretStr] = pydantic.Field(default=None, exclude=True)
-    private_key_path: Optional[str]
-    private_key_password: Optional[pydantic.SecretStr] = pydantic.Field(
-        default=None, exclude=True
-    )
-    authentication_type: Optional[str] = "DEFAULT_AUTHENTICATOR"
-    host_port: str
-    warehouse: Optional[str]
-    role: Optional[str]
-    include_table_lineage: Optional[bool] = True
-    include_view_lineage: Optional[bool] = True
-
-    connect_args: Optional[dict]
-
-    @pydantic.validator("authentication_type", always=True)
-    def authenticator_type_is_valid(cls, v, values, **kwargs):
-        valid_auth_types = {
-            "DEFAULT_AUTHENTICATOR": DEFAULT_AUTHENTICATOR,
-            "EXTERNAL_BROWSER_AUTHENTICATOR": EXTERNAL_BROWSER_AUTHENTICATOR,
-            "KEY_PAIR_AUTHENTICATOR": KEY_PAIR_AUTHENTICATOR,
-        }
-        if v not in valid_auth_types.keys():
-            raise ValueError(
-                f"unsupported authenticator type '{v}' was provided,"
-                f" use one of {list(valid_auth_types.keys())}"
-            )
-        else:
-            if v == "KEY_PAIR_AUTHENTICATOR":
-                # If we are using key pair auth, we need the private key path and password to be set
-                if values.get("private_key_path") is None:
-                    raise ValueError(
-                        f"'private_key_path' was none "
-                        f"but should be set when using {v} authentication"
-                    )
-                if values.get("private_key_password") is None:
-                    raise ValueError(
-                        f"'private_key_password' was none "
-                        f"but should be set when using {v} authentication"
-                    )
-            logger.info(f"using authenticator type '{v}'")
-        return valid_auth_types.get(v)
-
-    @pydantic.validator("include_view_lineage")
-    def validate_include_view_lineage(cls, v, values):
-        if not values.get("include_table_lineage") and v:
-            raise ValueError(
-                "include_table_lineage must be True for include_view_lineage to be set."
-            )
-        return v
-
-    def get_sql_alchemy_url(self, database=None):
-        return make_sqlalchemy_uri(
-            self.scheme,
-            self.username,
-            self.password.get_secret_value() if self.password else None,
-            self.host_port,
-            f'"{database}"' if database is not None else database,
-            uri_opts={
-                # Drop the options if value is None.
-                key: value
-                for (key, value) in {
-                    "authenticator": self.authentication_type,
-                    "warehouse": self.warehouse,
-                    "role": self.role,
-                    "application": APPLICATION_NAME,
-                }.items()
-                if value
-            },
-        )
-
-    def get_sql_alchemy_connect_args(self) -> dict:
-        if self.authentication_type != KEY_PAIR_AUTHENTICATOR:
-            return {}
-        if self.connect_args is None:
-            if self.private_key_path is None:
-                raise ValueError("missing required private key path to read key from")
-            if self.private_key_password is None:
-                raise ValueError("missing required private key password")
-            with open(self.private_key_path, "rb") as key:
-                p_key = serialization.load_pem_private_key(
-                    key.read(),
-                    password=self.private_key_password.get_secret_value().encode(),
-                    backend=default_backend(),
-                )
-
-            pkb = p_key.private_bytes(
-                encoding=serialization.Encoding.DER,
-                format=serialization.PrivateFormat.PKCS8,
-                encryption_algorithm=serialization.NoEncryption(),
-            )
-            self.connect_args = {"private_key": pkb}
-        return self.connect_args
-
-
-class SnowflakeConfig(BaseSnowflakeConfig, SQLAlchemyConfig):
-    database_pattern: AllowDenyPattern = AllowDenyPattern(
-        deny=[r"^UTIL_DB$", r"^SNOWFLAKE$", r"^SNOWFLAKE_SAMPLE_DATA$"]
-    )
-
-    database: Optional[str]  # deprecated
-
-    @pydantic.validator("database")
-    def note_database_opt_deprecation(cls, v, values, **kwargs):
-        logger.warning(
-            "snowflake's `database` option has been deprecated; use database_pattern instead"
-        )
-        values["database_pattern"].allow = f"^{v}$"
-        return None
-
-    def get_sql_alchemy_url(self, database: str = None) -> str:
-        return super().get_sql_alchemy_url(database=database)
-
-    def get_sql_alchemy_connect_args(self) -> dict:
-        return super().get_sql_alchemy_connect_args()
 
 
 class SnowflakeSource(SQLAlchemySource):
